// packages/ioc/src/handlers/interceptor.d.ts
import { Handler, HandlerFn } from './handler';
/**
 * Interceptor is a chainable behavior modifier for `hanlders`.
 *
 * 拦截器，用于链接多个处理器，组合成处理器串。
 */
export interface Interceptor<TInput = any, TOutput = any, TContext = any> {
    /**
     * the method to implemet interceptor.
     *
     * 实现拦截处理的方法
     * @param input  request input.
     * @param next The next handler in the chain, or the backend
     * if no interceptors remain in the chain.
     * @param context interceptor with context.
     * @returns An observable of the event stream.
     */
    intercept(input: TInput, next: Handler<any, TOutput, TContext>, context: TContext): TOutput;
    /**
     * is this equals to target or not
     *
     * 该实例等于目标与否？
     * @param target
     */
    equals?(target: any): boolean;
}
/**
 * interceptor fn.
 * 拦截方法，用于链接多个处理器，组合成处理器串。
 */
export type InterceptorFn<TInput = any, TOutput = any, TContext = any> = (input: TInput, next: HandlerFn<any, TOutput, TContext>, context: TContext) => TOutput;
/**
 * interceptor like.
 */
export type InterceptorLike<TInput = any, TOutput = any, TContext = any> = Interceptor<TInput, TOutput, TContext> | InterceptorFn<TInput, TOutput, TContext>;